backend/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "bmad-method"
version = "0.1.0"
description = "Serviços principais e ferramentas de suporte da plataforma Bmad Method"
requires-python = ">=3.10"
dependencies = [
  "pytest==8.4.2",
  "pytest-cov==7.0.0",
  "bandit==1.8.6",
  "pip-audit==2.9.0",
  "pyyaml==6.0.3",
  "fastapi==0.120.4",
  "uvicorn==0.38.0",
  "sqlalchemy==2.0.44",
  "pydantic==2.12.3",
  "strawberry-graphql==0.284.1",
  "httpx==0.28.1",
  "anyio==4.11.0",
  "psycopg[binary]==3.2.12",
  "opentelemetry-api==1.38.0",
  "opentelemetry-sdk==1.38.0",
  "opentelemetry-exporter-otlp==1.38.0",
  "opentelemetry-instrumentation-fastapi==0.59b0",
  "opentelemetry-instrumentation-sqlalchemy==0.59b0",
  "requests==2.32.5",
  "email-validator==2.3.0",
  "python-multipart==0.0.20",
  "pandas==2.3.3",
  "pactman==2.31.0",
]

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
include = [
  "services*",
  "quality*",
]
exclude = [
  "tests*",
]

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra"
testpaths = [
  "tests",
]
markers = [
  "unit: testes unitários de componentes isolados",
  "integration: testes de integração que validam orquestrações e requisitos de privacidade",
  "e2e: cenários ponta a ponta que simulam fluxos completos do MVP",
  "reservations: cenários críticos de reservas e gestão de disponibilidade",
  "payments: fluxos de tokenização, captura e reconciliação",
  "real_gateway: cenários que dependem do PSP real e são pulados quando o flag de contingência está desligado",
  "ota: contratos e integrações com parceiros OTA"
]
